CONFLUENT: CPKAFKA-1860; Send error response instead of failing connection (#32)
When the broker interceptor fails a request because it couldn't be transformed (e.g. invalid resource name or invalid principal that couldn't be prefixed), the interceptor throws an InvalidRequestException, which is propagated to the Processor processing the incoming receive. Processor treats this like any other invalid request/schema exception and closes the connection. For incoming requests which were successfully parsed but failed by the interceptor, this PR improves error handling by returning a response to the user with an INVALID_REQUEST error with the reason why the request was rejected.