CONFLUENT: CPKAFKA-1863: Workaround timing issue in Apache DS persistent search (#38)
As described in https://issues.apache.org/jira/browse/DIRSERVER-2257, there is a timing window in the persistent search handler in Apache DS because existing entries are processed before the persistent listener is created to process modifications. New entries created while existing entries are being processed, but before the persistent listener is created may never be returned to the client. This PR adds a workaround to ensure that missing entries are processed in the next iteration. Also adds error handling with retries during startup.