KAFKA-7304: Limit connection accept rate to prioritize other channels (#77)
Limit the number of new connections processed in each iteration in the SocketServer by
each Processor. Block the acceptor if the connection queue on all Processors are full.
Added a metric to track accept idle time percent.

This was originally committed to cc-kafka (reviewed by Jun, Jason and Dhruvil):

https://github.com/confluentinc/cc-kafka/commit/847553d23b60d1378f03750455f423c5ef015273

A couple of bugs were found and this commit includes fixes to both.

Reviewers: Ismael Juma <ismael@juma.me.uk>