CC-2978: Added support for free-tier and scheduled checking of licenses (#65)
Added support for detecting and using a free-tier license when the cluster(s) have only a single broker. LicenseManager always uses a single cluster for reading and storing its license, but other clusters can be checked to determine the free-tier applicability across multiple clusters.

Added `LicenseManager.start()` that will use a static instance of `LicenseManagers` (new) to continually check the licenses on a schedule. The calling component must register a listener to be notified of valid licenses expiring and expired licenses being renewed. This will allow the application to dynamically watch for updated licenses while running, and to optionally change its behavior when licenses expire.