CPKAFKA-1954: Enable CREATE_PARTITIONS request (#70)
We use the existing CreateTopicPolicy to validate the request explicitly in AdminManager.scala

This is a temporary change until KIP-201 is implemented. It is being driven by developers outside of Confluent. There is no use to implement our own KIP-201 because we would have to use the AK one in time.