CONFLUENT: Transform configs in CreateTopics request even when quotas aren't enabled (#119)
This patch fixes a bug with the CreateTopics logic in MultiTenantRequestContext. Previously, the logic to filter out non-updateable configs would not be run if quotas were not enabled. With this patch, configs will be filtered even without quotas enabled.
This patch also adds segment.ms to the list of updateable topic configs, to bring the code in line with the topic management documentation.

Reviewers: Stanislav, Rajini