Replace ce-trunk with master (#182)
We decided that we will use standard branch names with
multiple git repositories instead of prefixed branch names
in a single repository. The first step is to rename ce-trunk
to master.

Until we have defined the workflow, we will keep
cos-trunk in ce-kafka but we will rename it to ccs-master
since the cos term is no longer used and we want to use
`master` instead of `trunk` in CP projects. CCS stands
for Confluent Community Software.

Reviewers: Stanislav Kozlovski <stanislav_kozlovski@outlook.com>