CPKAFKA-2386: Kafka-topic-based metadata store for roles and groups (#176)
RBAC storage backend that uses a single writer and multiple readers. A single multi-partition topic is used to store RBAC roles and LDAP groups. Each broker has an embedded KafkaAuthStore instance that consumes all the entries from all partitions and stores the entries of the broker's scope in an in-memory cache, ensuring that there is no remote call in the authorize path for brokers.

Metadata Service is a broker plugin that uses the embedded KafkaAuthStore instance of the broker that hosts the service. This instance caches roles of all scopes to provide authorization for all components using the REST service. In addition to this, KafkaAuthStore of Metadata Service also supports updates using a single writer elected using Kafka group protocol.

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>, Ismael Juma <ismael@juma.me.uk>