CPKAFKA-2280: Tiered storage replication layer (#189)
Implements replication for tiered storage.

Adds two additional states to the replica fetcher: MaterializingTierMetadata and FetchingTierState. The fetcher thread will attempt to fetch from the leader, and upon hitting an OFFSET_TIERED error, will perform a TIER_LIST_OFFSETS request to retrieve the local disk offsets for the partition. It will then transition to MaterializingTierMetadata state, and will materialize tier metadata asynchronously until the metadata overlaps with the leaders local offsets. After doing so, it will fetch the tiered epoch state that aligns with the point that we will start replicating from, and will restore this epoch state. At any point, if we hit an error, we go back to the start, TIER_LIST_OFFSETS request -> MaterializingTierMetadata -> FetchingTierState. This approach has the advantage of replicating the hot set by always replicating the data that the leader has on disk.

The TierPartitionState now allows materialization of overlapping segment ranges, as long as they contain some additional data. The tier fetcher will currently preferentially fetch from the segment with the highest base offset.

As part of the replication tier state restore path a TierStateFetcher was added for use in fetching tier state such as the leader epoch state and the producer state snapshot from the object store. This is used by the replication code, and is primarily included so that state restoration has a separate queue from the TierFetcher, as we want this state to be restored as quickly as possible, and not be stuck behind large topic fetch requests.

Additional features:
- adds the ability to perform a TierListOffsetRequest to retrieve the leader's local log start and end offset
- adds kafka.tier.tools.DumpTierPartitionState to dump the contents of FileTierPartitionState as text.

Testing:

Adds several replication unit tests via mocking in AbstractFetcherThreadTest. Tests scenarios where tier state has to be restored, as well as scenarios where the local log is empty and no tier state restore is require.
TierEpochStateReplicationTest: two broker test where producer produces, follower is shutdown, producer continues producing, follower is started up, and the follower restores tier state.
TierEpochStateRevolvingReplicationTest: three broker test where a random follower is shutdown, producer produces, then the follower is started back up. Each restart requires the tier state to be restored. On each restart the tier epoch cache is compared between all brokers.

Co-authored-by: Dhruvil Shah <dhruvil@confluent.io>