[MINOR] Improve tier epoch replication test flakiness #239
As we are stopping brokers in these tests, there is a timing issue where the TierTopicManager doesn't manage to create the tier topic and doesn't become ready in time to archive the segments we produce. By the time it tries to create the tier topic, there are not enough brokers for the replication factor, meaning that all tiering is blocked.

Also improve logging when tier topic creation fails.