CONFLUENT: Accurately compute tierable log segments (#228)
We have been a bit conservative about finding tierable segments. We always exclude the last candidate segment, as it might contain offsets that are not tierable. This patch makes this logic more accurate. We now have logic to figure out whether such untierable offsets are indeed in the last candidate segment.