CPKAFKA-2665; Callback handler for LDAP username/password authentication (#256)
Callback handler for Metadata Service to perform LDAP username/password authentication for clients using HTTP Basic authentication method. This uses the Kafka AuthenticateCallbackHandler interface so that it can also be used in future for SASL/PLAIN authentication using LDAP for Kafka clients when custom configs are supported for callback handlers.

To avoid external dependencies, password comparison is performed using an extensible interface. Encrypted password verification based on the implementation in Jetty is included in Metadata Service.

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>